Various fixes for Modals (#326)
* Resync with master (#9)

* Create form-input-static.vue

New form-static input

* Added form-input-static compoinent

* Refactored static input

Refactored to use the new `<b-form-input-static>` component

* Switch to bFormInputStatic

Updated child component var to bFormInputStatic to follow proper naming conventions

* Removed lazyFormatter from static-input

* Added trailing semi-colon

To make CircleCI happy

* Added missing 'this'

* [nav-item] add dropdown class

* Added <slot> for robustness

* new b-form-input-static component (#292)

* Create form-input-static.vue

New form-static input

* Added form-input-static compoinent

* Refactored static input

Refactored to use the new `<b-form-input-static>` component

* Switch to bFormInputStatic

Updated child component var to bFormInputStatic to follow proper naming conventions

* Removed lazyFormatter from static-input

* Added trailing semi-colon

To make CircleCI happy

* Added missing 'this'

* Added <slot> for robustness

* fixed missing `.vue` extension on import

* Added missing extension on component import (#293)

* Optimized import order in form-input.vue (#294)

* Added missing extension on component import

* Optimized import order

* Refactoring focus timing

* Update to include modal

* Ensure focusable item if focused when modal opens

Ensures that the first focusable item in the modal is focused when it is shown.
When selecting the item to focus:
- First looks in footer
- Then body
- Then header
- Else focus the modal itself

Also included generate-id mixin for ARIA labeling

* Added missing space

* Fixed Typo and referenced _id on $emit/$on

* fixed focusin event

Changed from listing for focus to focusin for enforceFocus hanlder

* Moved add/removeEventlistener to show()/hide()

enforceFocus event listeners are now only instantiated when the modal is shown.

* Minor updates to focus handler

* Feature to return focus to trigger element

Added a second optional argument to the `show::modal` event to allow specifying the element (i.e. button, link) to return focus to when Modal closes.

* Allowed passing original trigger element

Allow to return focus to triggering element when modal is closed.

* Allow for prop to specify return focus element

added in prop `returnFocus` to specify element to re-focus to when modal closes.

Accepts either a CSS selector string (i.e. '#mybutton') or an `HTMLElement` reference. If CSS selector string matches more than one element, then only the first element is re-focused.

* Removed auto_id generation

`aria-labelledby` and `aria-describedby` will not be present (by setting bound attribute to null) if an `id` is not specified for the modal.

* Removed reference to modal.vue