feat(form-select): Allow selectSize to be set when not in multiple mode (Issue #761) (#762)
* [form-select] Allow select-size to be set when not in multiple mode

Allows the select native `size` attribute to be set (via the prop `select-size`), even when not in multiple mode.  Only has any effect when `select-size > 1`.  custom select styling is removed when in `multiple` mode or `select-size > 1`.

Changes `multiple-size` attribute to `select-size` (`multiple-size` was not available in v0.18.0, so deprecation is not needed)

* [form-select] update documentation