feat(table): add responsive stacked table option
This adds a new prop `stacked`, which allows the table to render in a visually stacked format based on breakpoint. (CSS assumes standard bootstrap breakpoints).

If no value is passed to `stacked`, the table will always be rendered as stacked.

If `stacked` is set to a breakpoint of `sm`, `md`, `lg`, or `xl`, then the table will only appear stacked at view ports below the specified breakpoint.

When visually stacked, the table thead and tfoot will not be shown.  Cells will be prefixed with the fields label (not the scoped HEAD slot). Because of this, it is not possible to click the "headings" to sort the columns.  Users however, can provide an external control (such as b-form-select) to control the sorting and sort direction (via b-table's `sort-by` and `sort-desc` syncable props).