Ignore BV keys when auto-discovering table fields
Bootstrap-Vue has some extra keys that can be provided to enhance table
functionality. These should be explicitly ignored when automatically
discovering the fields for a table, when the fields prop is not
provided.

The reason for explicitly ignoring the keys, rather than blanket
ignoring all keys beginning with "_" is to minimise any possible
collisions with user provided data in that field.

Fixes #1635