fix(select): Fix issues with form-select
The form-input component currently has some arbitary restrictions around
the combinations of multiple, custom-select and selectSize that will be
applied. Previously I tried to address this by passing through the
selectSize, but this caused an issue (hiding the arrows) in custom
selects when the size was set to 0.

This patch cleans up the internal computed state of form-select to be
more relaxed about what user input is rendered in the final component

Fixes: #1658