feat(b-table, b-table-lite): place `<tfoot>` after `<tbody>` element per HTML5 spec
In HTML4 spec, the `<tfoot>` element must be placed **before** the `<tbody>`, while HTML5 spec states that is must appear **after** the `<tbody>` element.  

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/tfoot


> A <table> element. The `<tfoot>` must appear after any `<caption>`, `<colgroup>`, `<thead>`, `<tbody>`, or `<tr>` element. Note that this is the requirement as of HTML5.
> > HTML 4 The `<tfoot>` element cannot be placed after any `<tbody>` and `<tr>` element. Note that this directly contradicts the above normative requirement from HTML5.

Placing the `<tfoot>` after the `<tbody>` also provides better semantic markup flow, specifically for screen readers when a table display has been changed to `display: block;` via CSS (which removes the normal semantic meaning of table elements). 

ARIA roles can help (which we do render on all table elements), but there is no diferentiation between `thead` or `tfoot` via the `rowgroup` role. `tfoot` and `thead` cells have the role `columnheader`, but again this does not differentiate a header cell vs a footer cell.

By moving the `<tfoot>` to the bottom of the rendered table, more semantic meaning can be derived solely based on the position of the `<tfoot>` within the `<table>`.