chore(b-form-file): Fix types allowed by value (prop type check)
Allow empty string to be passed.

When creating the new props docs and web-types (not yet released), the value type was changed to `[File, Array]` from Any (`undefined`).

The resetting of the file control can be done by setting to an empty string, and when doing so Vue would emit an prop type check warning.  This adds back in the type String to prevent the warning.